app_v0/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy>=2.0
psycopg[binary]
python-dotenv
streamlit
httpx
